requirements.txt
black == 23.3.0
boto3 == 1.26.155
fastapi == 0.96.0
pandas == 2.0.2
poetry==1.5.1
pydantic == 1.10.9
psycopg2-binary == 2.9.6 
python-dotenv == 1.0.0
pytest == 7.4.0
requests == 2.31.0
SQLAlchemy == 2.0.15
uvicorn[standard] == 0.22.0